practica

+++ app/Http/Controllers/BoardController.php
@@ -4,15 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Board;
 use App\Models\BoardUser;
-use App\Models\User;
 use App\Models\Task;
+use App\Models\User;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Http\Request;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Facades\Redirect;
 
 /**
  * Class BoardController
@@ -43,47 +45,16 @@ class BoardController extends Controller
         }
 
         $boards = $boards->paginate(10);
-        $users = User::all();
 
         return view(
             'boards.index',
+            'dashboard.index',
             [
                 'boards' => $boards,
-                'users' => $users
+                'userList' => User::select(['id', 'name'])->get()->toArray()
             ]
         );
     }
-
-    public function updateBoard(Request $request, $id): JsonResponse
-    {
-        $board = Board::find($id);
-
-        $error = '';
-        $success = '';
-
-        if ($board) {
-
-            $board->boardUsers()->delete();
-
-            $board_user_ids = $request->get('board_user_ids');
-
-            if (isset($board_user_ids)) {
-                foreach ($board_user_ids as $board_user_id) {
-                    BoardUser::query()->create([
-                        'user_id' => $board_user_id,
-                        'board_id' => $id
-                    ]);
-                }
-            }
-
-            $success = 'Board saved';
-        } else {
-            $error = 'Board not found!';
-        }
-
-        return response()->json(['error' => $error, 'success' => $success, 'board' => $board]);
-    }
-
 
     /**
      * @param  Request  $request
@@ -91,19 +62,80 @@ class BoardController extends Controller
      *
      * @return JsonResponse
      */
-    public function deleteBoard(Request $request, $id): JsonResponse
+    public function updateBoard(Request $request, $id): JsonResponse
     {
+        /** @var Board $board */
         $board = Board::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
 
         $error = '';
         $success = '';
 
         if ($board) {
-            $board->delete();
+            if ($board->user->id === $user->id || $user->role === User::ROLE_ADMIN) {
+                $newBoardUsers = $request->get('boardUsers');
+                $existingBoardUsers = [];
 
-            $success = 'User deleted';
+                $board->boardUsers()->get()->each(function ($boardUser) use ($newBoardUsers, &$existingBoardUsers) {
+                    if (!in_array($boardUser->user_id, $newBoardUsers)) {
+                        $boardUser->delete();
+                    } else {
+                        $existingBoardUsers[] = $boardUser->user_id;
+                    }
+                });
+
+                $toSave = array_diff($newBoardUsers, $existingBoardUsers);
+
+                foreach ($toSave as $userId) {
+                    $boardUser = new BoardUser();
+                    $boardUser->board_id = $board->id;
+                    $boardUser->user_id = $userId;
+                    $boardUser->save();
+                }
+
+                $board->name = $request->get('name');
+                $board->save();
+                $board->refresh();
+
+                $success = 'Board saved';
+            } else {
+                $error = 'You don\'t have permission to edit this board!';
+            }
         } else {
-            $error = 'User not found!';
+            $error = 'Board not found!';
+        }
+
+        return response()->json(['error' => $error, 'success' => $success, 'board' => $board]);
+    }
+
+    /**
+     * @param $id
+     *
+     * @return JsonResponse
+     */
+    public function deleteBoard($id): JsonResponse
+    {
+        /** @var Board $board */
+        $board = Board::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        $error = '';
+        $success = '';
+
+        if ($board) {
+            if ($board->user->id === $user->id || $user->role === User::ROLE_ADMIN) {
+                $board->delete();
+
+                $success = 'Board deleted';
+            } else {
+                $error = 'You don\'t have permission to delete this board!';
+            }
+        } else {
+            $error = 'Board not found!';
         }
 
         return response()->json(['error' => $error, 'success' => $success]);
@@ -139,15 +171,172 @@ class BoardController extends Controller
             return redirect()->route('boards.all');
         }
 
-        $tasks = Task::query()->where('board_id', $id)->paginate(10);
+        $tasks = $board->tasks()->oldest()->paginate(10);
+
+        $boardUsers = $board->boardUsers()->with('user')->get();
 
         return view(
             'boards.view',
             [
                 'board' => $board,
                 'boards' => $boards,
-                'tasks' => $tasks
+                'tasks' => $tasks,
+                'boardUsers' => $boardUsers
             ]
         );
+    }
+
+
+    public function createBoard(Request $request)
+    {
+        $error = '';
+        $success = '';
+
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'boardUsers' => 'array'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()]);
+        }
+
+
+        $board = Board::create([
+            'user_id' => Auth::id(),
+            'name' => $request->get('name')
+        ]);
+
+        $userIds = $request->get('boardUsers');
+
+        foreach ($userIds as $userId) {
+            $boardUser = new BoardUser();
+            $boardUser->board_id = $board->id;
+            $boardUser->user_id = $userId;
+            $boardUser->save();
+        }
+
+        return response()->json(['error' => $error, 'success' => $success]);
+
+        return redirect(route('boards.all'));
+    }
+
+    public function createTask(Request $request)
+    {
+        $error = '';
+        $success = '';
+
+        $validator = Validator::make($request->all(), [
+            'name' => 'required',
+            'description' => 'required',
+            'assignment' => 'required',
+            'status' => 'required',
+            'board_id' => 'required'
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['error' => $validator->errors()]);
+        }
+
+
+        $board = Board::find($request->get('board_id'));
+
+        if (!$board) {
+            return response()->json(['error' => ['board_id' => 'No board']]);
+        }
+
+        $task = Task::create([
+            'name' => $request->get('name'),
+            'board_id' => $board->id,
+            'description' => $request->get('description'),
+            'assignment' => $request->get('assignment'),
+            'status' => $request->get('status')
+        ]);
+
+
+        if ($task) {
+            $task->name = $request->get('name');
+            $task->description = $request->get('description');
+            $task->assignment = $request->get('assignment');
+            $task->status = $request->get('status');
+            $task->save();
+
+            $success = 'Task saved';
+        } else {
+            $error = 'Task hasn\'t been saved';
+        }
+
+
+        return response()->json(['error' => $error, 'success' => $success]);
+    }
+
+
+    /**
+     * @param  Request  $request
+     * @param $id
+     *
+     * @return JsonResponse
+     */
+    public function updateTask(Request $request, $id): JsonResponse
+    {
+        /** @var Task $task */
+        $task = Task::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        $boardUser = BoardUser::where('board_id', $task->board_id)->where('user_id', $user->id)->first();
+
+        $error = '';
+        $success = '';
+
+        if ($task) {
+            if ($boardUser) {
+                $task->name = $request->get('name');
+                $task->description = $request->get('description');
+                $task->assignment = $request->get('assignment');
+                $task->status = $request->get('status');
+                $task->save();
+
+                $success = 'Task saved';
+            } else {
+                $error = 'You don\'t have permission to edit this task!';
+            }
+        } else {
+            $error = 'Task not found!';
+        }
+
+        return response()->json(['error' => $error, 'success' => $success, 'task' => $task]);
+    }
+
+    /**
+     * @param $id
+     *
+     * @return JsonResponse
+     */
+    public function deleteTask($id): JsonResponse
+    {
+        /** @var Task $task */
+        $task = Task::find($id);
+
+        /** @var User $user */
+        $user = Auth::user();
+
+        $error = '';
+        $success = '';
+
+        if ($task) {
+            if ($task->board->user->id === $user->id || $user->role === User::ROLE_ADMIN) {
+                $task->delete();
+
+                $success = 'Task deleted';
+            } else {
+                $error = 'You don\'t have permission to delete this task!';
+            }
+        } else {
+            $error = 'Task not found!';
+        }
+
+        return response()->json(['error' => $error, 'success' => $success]);
     }
 }
